Fix bug where Player was moving back to beginning on scene reload
This is because on the first frame the scene loads, it calculates a direction based on the camera position (line 83, Player.cs). It transforms this direction from local to world space and since when the scene first loads, the camera is pointed at the very beginning, it uses the position from the beginning and moves the player back there, even though there is a save point.
The fix is to just check if there's no player input, then just return and don't calculate the movement